Add Drive module

+++ CppSynth/Synth.cpp
@@ -404,7 +404,7 @@ namespace Leiftur
 	void Synth::SetParameterInner(Module module, int parameter, double value)
 	{
 		auto paramInfo = Parameters::ParamInfo[module][parameter];
-		value = Utils::Limit(value, paramInfo.MinValue, paramInfo.MaxValue);
+		value = Utils::Limit((float)value, (float)paramInfo.MinValue, (float)paramInfo.MaxValue);
 		int idx = PackParameter(module, parameter);
 		currentPreset.Values[idx] = value;
 		SendBackParameter(module, parameter);
